+++ src/main/scala/uk/co/gresearch/spark/package.scala
@@ -27,10 +27,7 @@ import org.apache.spark.storage.StorageLevel
 import uk.co.gresearch.spark.group.SortedGroupByDataset
 import uk.co.gresearch.spark._
 
-package object spark extends Logging with SparkVersion {
-
-  val SparkCompatVersion: (Int, Int) = (SparkCompatMajorVersion, SparkCompatMinorVersion)
-  val SparkCompatVersionString: String = s"$SparkCompatMajorVersion.$SparkCompatMinorVersion"
+package object spark extends Logging with SparkVersion with BuildVersion {
 
   /**
    * Provides a prefix that makes any string distinct w.r.t. the given strings.