+++ src/test/scala/uk/co/gresearch/spark/SparkSuite.scala
@@ -29,7 +29,7 @@ import uk.co.gresearch.spark.SparkSuite.Value
 import java.sql.Timestamp
 import java.time.Instant
 
-class SparkSuite extends AnyFunSuite with SparkTestSession with SparkVersion {
+class SparkSuite extends AnyFunSuite with SparkTestSession {
 
   import spark.implicits._
 
@@ -37,9 +37,15 @@ class SparkSuite extends AnyFunSuite with SparkTestSession with SparkVersion {
   val emptyDataFrame: DataFrame = spark.createDataFrame(Seq.empty[Value])
 
   test("Get Spark version") {
-    assert(spark.version.startsWith(s"$SparkCompatMajorVersion.$SparkCompatMinorVersion."))
-    assert(SparkCompatVersion === (SparkCompatMajorVersion, SparkCompatMinorVersion))
-    assert(SparkCompatVersionString === s"$SparkCompatMajorVersion.$SparkCompatMinorVersion")
+    assert(VersionString.contains(s"-$BuildSparkCompatVersionString-") || VersionString.endsWith(s"-$BuildSparkCompatVersionString"))
+
+    assert(spark.version.startsWith(s"$BuildSparkCompatVersionString."))
+    assert(SparkVersion === BuildSparkVersion)
+    assert(SparkCompatVersion === BuildSparkCompatVersion)
+    assert(SparkCompatVersionString === BuildSparkCompatVersionString)
+    assert(SparkMajorVersion === BuildSparkMajorVersion)
+    assert(SparkMinorVersion === BuildSparkMinorVersion)
+    assert(SparkPatchVersion === BuildSparkPatchVersion)
   }
 
   Seq(MEMORY_AND_DISK, MEMORY_ONLY, DISK_ONLY, NONE).foreach { level =>